% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_crab_data.R
\name{clean_crabs_data}
\alias{clean_crabs_data}
\title{Omit NA Values}
\usage{
clean_crabs_data(data)
}
\arguments{
\item{data}{.csv file as a data frame}
}
\value{
data A data frame that will have to be saved as a new data frame (numeric)
}
\description{
Takes a .csv saved as a data frame and removes any NA values. 
Returns the data frame with no NA values.
}
