% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crab_area.R
\name{crab_area}
\alias{crab_area}
\title{Measures area of carapace}
\usage{
crab_area(data, column1, column2)
}
\arguments{
\item{data}{.csv file as a data frame}

\item{column1}{numerical column of length measurements}

\item{column2}{numerical column of width measurements}
}
\value{
area Value from column 1 multiplied by column 2 (numeric)
}
\description{
Takes the length and width of the carapace of the crab and multiplies them. 
Returns the area of the carapace.
}
