% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_crabs.R
\name{graph_crabs}
\alias{graph_crabs}
\title{Plots a point graph}
\usage{
graph_crabs(df, x, y)
}
\arguments{
\item{data}{.csv file as a data frame}

\item{column1}{numerical column of length measurements}

\item{column2}{numerical column of width measurements}
}
\value{
area Value from column 1 multiplied by column 2 (numeric)
}
\description{
Takes two separate columns of a data frame.
Returns a point graph.
}
