% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_m_unit_conversion.R
\name{mm_m_unit_change}
\alias{mm_m_unit_change}
\title{Converts the unit millimeters to meter}
\usage{
mm_m_unit_change(data, col1)
}
\arguments{
\item{data}{.csv file as a data frame}

\item{col1}{numerical column of measurement that needs converting}
}
\value{
meters The same measurement as a meter (numeric)
}
\description{
Takes a measurement of millimeters usually used to measure smaller crabs and divides by 1000. 
Returns the same measurement meters.
}
