% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crabs.R
\name{plot_crabs}
\alias{plot_crabs}
\title{Measures area of carapace}
\usage{
plot_crabs(data, col_dep, col_indep)
}
\arguments{
\item{data}{.csv file as a data frame}

\item{col_dep}{a column of data that serves as the dependent variable}

\item{col_indep}{a different column of data that serves as the independent variable}
}
\value{
linear data output
}
\description{
Takes the length and width of the carapace of the crab and multiplies them. 
Returns the area of the carapace.
}
